Use html <table> instead of <pre> for large output commands.
While still malformed on Mumla when there are multiple columns there
are at least some line breaks to make things a bit nicer.

Additionally, the main mumble client respects the chatbox width with
tables but not with <pre> so this should reduce the amount of times
the bot will randomly expand the chatbox and cause a misalignment.