Add Summon command which also re-targets users, should fix #16
While it would make far more sense to retarget users (and thus channel)
during the ChannelChange event I have found this to be unreliable
when moving the bot via another user dragging the bot via mumble.

Instead we implement a summon command which can be messaged to the bot
where the bot will attempt to join the channel in a standardized way
rather than relying on server powers.