Fix: joinUserChannel joins channel directly rather than failing
Previously we used the Channels.Find() method improperly which often
lead to the target channel being nil, aside from not having a necessary
nil check this often (or always) failed to find the channel we intended
to join leading to a null pointer exception.

Instead, we now move to the channel directly and have an additional
check in place to ensure that the channel we intend to move to is not
nil.