Refactor play() to reduce nested ifs and improve readability
I believe this maintains all of the previous functions and behaviors
with far less of an if statement mess.

We adopt player.PlayCurrent() to add an additional check to not start
if the player is already playing as this is almost always non-desired
behavior and we will deliberately stop playback before playing again.