Only accept bare commands in private messages, fixes #24
Previously, Username was never actually set in the configuration leading
to IsCommand always being true as HasPrefix(s, "") returns true always.

Now, we properly set the username during connect, via a flag lookup.

Currently we're taking the username out of the configuration block as
that was never really used for anything outside of IsCommand.