feat: read URL prefix whitelist from `whitelist.txt` for yt-dlp
whitelist.txt is loaded at STARTUP and does not currently check for
changes to to the file.

We should now be able to load a list of URL prefixes (one per line)
into the bot at startup to add more flexibility than the previously
hardcoded list of youtube and soundcloud URLs

Advice about ending the prefix with / still applies and is enforced,
in the future we may wish for a more granular control of the whitelist,
for instance we can't allow a single URL given our enforcement of ending
with "/".